Remove fluvio-system-util crate
