--- crates/fluvio-system-util/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "fluvio-system-util"
-version = "0.1.1"
-authors = ["fluvio.io"]
-edition = "2018"
-description = "Utilities for Fluvio System components"
-repository = "https://github.com/infinyon/fluvio"
-license = "Apache-2.0"
-
-[features]
-default = []
-fixture = []
-
-[dependencies]
-tracing = "0.1.19"
-rand = "0.7.2"
-
-# Fluvio dependencies
-fluvio-types = { version = "0.2.0", path = "../fluvio-types"  }
