--- crates/fluvio-system-util/src/counters/mod.rs
@@ -1,6 +1,0 @@
-mod counter_tbl;
-#[allow(clippy::module_inception)]
-mod counters;
-
-pub use self::counter_tbl::CounterTable;
-pub use self::counters::Counters;
