--- crates/fluvio-system-util/src/lib.rs
@@ -1,4 +1,0 @@
-pub mod config_helper;
-pub mod counters;
-pub mod generators;
-pub mod bin;
